pyproject.toml
[tool.poetry]
name = "detextify"
version = "0.1.0"
description = ""
authors = ["Julia <turc.raluca@gmail.com>", "Mihail <mihail911@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.6.0.66"
numpy = "^1.23.5"
absl-py = "^1.3.0"
azure-cognitiveservices-vision-computervision = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
